+++ forgerock-openbanking-aspsp/forgerock-openbanking-aspsp-rs/forgerock-openbanking-rs-api/src/main/java/com/forgerock/openbanking/aspsp/rs/ext/lbg/file/payment/csv/api/v3_1_3/CSVFilePaymentConsentsApiController.java
@@ -67,7 +67,7 @@ import static com.forgerock.openbanking.constants.OpenBankingConstants.HTTP_DATE
 
 @Api(value = "csv-file-payment-consents", description = "the CSV file-payment-consents API")
 @Controller("CSVFilePaymentConsentsApiV3.1.3")
-@RequestMapping("/open-banking/v3.1.3/pisp")
+@RequestMapping({"/open-banking/v3.1.3/pisp","/open-banking/v3.1.4/pisp"})
 @Slf4j
 public class CSVFilePaymentConsentsApiController {
