+++ pom.xml
@@ -50,7 +50,7 @@
         <ob-auth.version>1.0.57</ob-auth.version>
         <ob-directory.version>1.4.73</ob-directory.version>
         <ob-auth.version>1.0.57</ob-auth.version>
-        <ob-aspsp.version>1.0.87</ob-aspsp.version>
+        <ob-aspsp.version>1.0.88</ob-aspsp.version>
         <ob-clients.version>1.0.34</ob-clients.version>
         <ob-extensions.version>1.0.10</ob-extensions.version>
     </properties>
